auth_api/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authorization service"
authors = ["IAmIrina, alexshvedov1997"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
gevent = "^21.12.0"
psycopg2-binary = "^2.9.3"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
Flask-SQLAlchemy = "^2.5.1"
redis = "^4.3.4"
gunicorn = "^20.1.0"
Flask-Migrate = "^3.1.0"
Flask-RESTful = "0.3.9"
email-validator = "1.2.1"
passlib = "1.7.4"
Flask-JWT-Extended = "^4.4.3"
orjson = "^3.7.11"
flask-marshmallow = "0.14.0"
webargs = "8.2.0"
marshmallow-sqlalchemy = "0.28.1"
flask-rest-paginate = "^1.1.3"
flasgger = "0.9.5"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
flake8-isort = "^4.2.0"
autopep8 = "^1.7.0"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
